requirements.txt
flask
flask-cors
docker
psutil
websocket-client